+++ changelog.md
@@ -17,6 +17,7 @@
 * Ajout d'une page tableau récap pour la veille "tableau_hier.php"
 * Ajout en paramètres de nom et de l'URL de l'asso qui apparait dans les crédits des graphiques
 * Ajout de quelques paramètres supplémentaires dans le makejson_48h
+* Ajout de graphiques highcharts sur 48 heures (Température-Humidité, Pression, Précipitations, UV, Rayonnement solaire)
 
 ### 0.3 - 29/10/2016
 * Ajout de paramètres UV et rayonnement dans le fichier de conf, selon si la station est équipée de la sonde ou non